Update agent formula to choose an m1/intel tarball based on the host computer
Homebrew is working through the complexity of transitioning from Intel
to Apple Silicon and much is up in the air.

However, our agent formula is quite simple. On Apple Silicon machines,
we can ignore whether homebrew itself is running natively and just
always download the Apple Silicon version of the agent.

On non Apple Silicon machines, we can ignore whether homebrew itself is
running natively and just always download the Intel version of the
agent.

This requires some helpers added to homebrew in [1]. Will they be
available everywhere?

Before merging this we also need to consider the agent CD scripts that
update this file automatically [2].

[1] https://github.com/Homebrew/brew/pull/7995/files
[2] https://github.com/buildkite/agent/blob/b2cf4e3d90f5b155505ec6eb53e007a4d88239d7/.buildkite/steps/release-homebrew.sh